+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## Unreleased ()
+
+* Add InfluxDB notifier type.
+* Fix bug where only the first notifier would be called if the OVF module included more than one test.
+
 ## 1.1.0 (2016-09-03)
 
 * Allow notifiers to be optionally executed upon successful tests.